Move plugin tests that don't use plugin to test_main_widget.py
